Move the code in views to view models.
