+++ M04/Assignment/SprintPokemonChallenge/app/src/main/AndroidManifest.xml
@@ -12,9 +12,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.PokemonDetailsActivity"></activity>
-        <activity android:name=".ui.ListOfPokemon" />
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".view.PokemonDetailsActivity"></activity>
+        <activity android:name=".view.ListOfPokemon" />
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
