+++ M04/Assignment/SprintPokemonChallenge/app/src/main/res/layout/activity_main.xml
@@ -1,53 +1,56 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.MainActivity">
+    xmlns:tools="http://schemas.android.com/tools">
 
-    <EditText
-        android:id="@+id/edit_txt"
-        android:layout_margin="4dp"
-        app:layout_constraintHorizontal_weight="7"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:hint="@string/search_pokemon_id_or_name"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toLeftOf="@id/btn_search"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="@+id/btn_search"/>
+    <data>
+        <variable
+            name="viewModel"
+            type="com.example.sprintpokemonchallenge.viewmodel.MainViewModel" />
+        <variable
+            name="main"
+            type="com.example.sprintpokemonchallenge.view.MainActivity" />
+    </data>
 
-    <Button
-        android:id="@+id/btn_search"
-        android:drawableTop="@drawable/big_search"
-        android:layout_margin="4dp"
-        app:layout_constraintHorizontal_weight="1"
-        android:layout_width="0dp"
-        android:layout_height="50dp"
-        app:layout_constraintLeft_toRightOf="@id/edit_txt"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler"
-        android:layout_margin="4dp"
-        android:clickable="true"
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:listitem="@layout/recycler_item"
-        app:layout_constraintTop_toBottomOf="@id/edit_txt"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
+        android:layout_height="match_parent"
+        tools:context=".view.MainActivity">
 
-<!--    <Button-->
-<!--        android:id="@+id/btn_list"-->
-<!--        android:layout_margin="8dp"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:text="@string/my_pokemon_list"-->
-<!--        app:layout_constraintBottom_toBottomOf="parent"-->
-<!--        app:layout_constraintLeft_toLeftOf="parent"-->
-<!--        app:layout_constraintRight_toRightOf="parent" />-->
+        <EditText
+            android:id="@+id/edit_txt"
+            android:layout_margin="4dp"
+            app:layout_constraintHorizontal_weight="7"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:hint="@string/search_pokemon_id_or_name"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toLeftOf="@id/btn_search"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="@+id/btn_search" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <Button
+            android:id="@+id/btn_search"
+            android:drawableTop="@drawable/big_search"
+            android:layout_margin="4dp"
+            app:layout_constraintHorizontal_weight="1"
+            android:layout_width="0dp"
+            android:layout_height="50dp"
+            app:layout_constraintLeft_toRightOf="@id/edit_txt"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:onClick="@{(view) -> viewModel.onClickSearch(main.editText, main.main)}"/>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recycler"
+            android:layout_margin="4dp"
+            android:clickable="true"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            tools:listitem="@layout/recycler_item"
+            app:layout_constraintTop_toBottomOf="@id/edit_txt"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
